+++ scripts/qso_months.py
@@ -12,6 +12,7 @@ import logging.config
 import matplotlib.gridspec as gridspec
 from qry import *
 from plotuty import saveplt
+from calendar import monthrange
 
 logging.config.fileConfig("../cfg/log_config.ini", disable_existing_loggers=False)
 file_output = '../static/plots/'+ os.path.splitext(os.path.basename(sys.argv[0]))[0]
@@ -24,10 +25,12 @@ qry_string="""
 
 	select 
 		concat(YEAR(FROM_UNIXTIME(time)),'-',  
-		MONTH(FROM_UNIXTIME(time))) as ym, 
+		right(concat('0',MONTH(FROM_UNIXTIME(time))),2)
+                ) as ym, 
 		count(0) as number
 		from spot 
-		GROUP by 1;
+		GROUP by 1
+                ORDER by 1;
 
     """
 logging.debug(qry_string) 
@@ -56,6 +59,15 @@ plt.grid(False)
 plt.subplots_adjust(left=0.15)
 
 plt.bar(x, y, align='center')
+list_y = list(y)
+le=len(list_y)-1
+day=datetime.today().day
+month=datetime.today().month
+year=datetime.today().year
+days_of_month=monthrange(year,month)
+list_y[le]=int(y[le]/day*days_of_month[1])
+y = tuple(list_y)
+plt.scatter(x,y)
 
 saveplt(plt,file_output)