+++ docs/CHANGELOG.md
@@ -10,6 +10,9 @@ xx/xx/2020: 2.0
 - removed link to country-flags.io and used https://github.com/lipis/flag-icon-css hosted at https://cdnjs.com/libraries/flag-icon-css/3.4.1 in order to improve performances and remove cookies
 - removed cookie bar
 - improved offline page
+- remove colors from callsing search resoults table
+- added popover on mobile
+- added monitor.sh in order to monitor your system
 ___
 04/10/2020: 1.2.2
 - Added config.sh utilty in script folder, in order to avoid edit manually config.json